- Remove minion-identity bundle and instead split out the rest client as a separate bundle - Share the controller config with th rest client
grpc scv impl depends on core-impl
rest client depends on grpc scv impl and on core-impl

These have to be separate to avoid circular deps
