+++ features/distributed/core/api/src/main/java/org/opennms/distributed/core/api/ControllerConfig.java
@@ -38,4 +38,6 @@ public interface ControllerConfig {
 
     int getBrokerIdleTimeout();
 
+    String getHttpUrl();
+
 }
