+++ features/minion/core/features/src/main/resources/features.xml
@@ -11,16 +11,8 @@
         <feature>camel-core</feature>
         <feature>camel-blueprint</feature>
         <feature>camel-jaxb</feature>
-        <feature>opennms-health</feature>
         <feature>opennms-distributed-core-impl</feature>
-        <feature>scv-api</feature>
 
-        <!-- Used for the REST client -->
-        <bundle>mvn:org.apache.httpcomponents/httpcore-osgi/${httpcoreVersion}</bundle>
-        <bundle>mvn:org.apache.httpcomponents/httpclient-osgi/${httpclientVersion}</bundle>
-        <bundle>mvn:com.google.code.gson/gson/${gsonVersion}</bundle>
-
-        <bundle>mvn:org.opennms.core.health/org.opennms.core.health.api/${project.version}</bundle>
         <bundle>mvn:org.opennms.features.minion/core-impl/${project.version}</bundle>
     </feature>
