+++ features/minion/core/impl/src/main/java/org/opennms/minion/core/impl/MinionIdCommand.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.minion.core.identity;
+package org.opennms.minion.core.impl;
 
 import org.apache.karaf.shell.api.action.Command;
 import org.apache.karaf.shell.api.action.lifecycle.Service;
