+++ features/minion/core/impl/src/main/java/org/opennms/minion/core/impl/MinionIdentityImpl.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.minion.core.identity;
+package org.opennms.minion.core.impl;
 
 import org.opennms.distributed.core.api.MinionIdentity;
 import org.opennms.distributed.core.api.SystemType;
