--- features/minion/core/minion-identity-impl/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,30 +1,0 @@
-<blueprint
-        xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
-        xsi:schemaLocation="
-        http://www.osgi.org/xmlns/blueprint/v1.0.0 
-        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-
-        http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
-        http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
-">
-    <cm:property-placeholder id="minionProperties" persistent-id="org.opennms.minion.controller" placeholder-prefix="[[" placeholder-suffix="]]" update-strategy="reload">
-        <cm:default-properties>
-            <cm:property name="location" value="MINION"/>
-            <cm:property name="id" value="00000000-0000-0000-0000-000000ddba11"/>
-        </cm:default-properties>
-    </cm:property-placeholder>
-
-    <!-- A canonical source for this Minion's identity -->
-    <bean id="minionIdentity" class="org.opennms.minion.core.identity.MinionIdentityImpl">
-        <argument value="[[location]]"/>
-        <argument value="[[id]]"/>
-    </bean>
-    <service ref="minionIdentity" interface="org.opennms.distributed.core.api.MinionIdentity"/>
-    <service ref="minionIdentity" interface="org.opennms.distributed.core.api.Identity">
-        <service-properties>
-            <entry key="systemType" value="Minion"/>
-        </service-properties>
-    </service>
-</blueprint>
