+++ features/minion/core/minion-rest-client/pom.xml
@@ -8,9 +8,9 @@
         <version>27.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>minion-identity-impl</artifactId>
+    <artifactId>minion-rest-client</artifactId>
     <packaging>bundle</packaging>
-    <name>OpenNMS :: Features :: Minion :: Core :: Identity Impl</name>
+    <name>OpenNMS :: Features :: Minion :: ReST Client</name>
 
     <build>
         <plugins>
@@ -42,10 +42,18 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.features.distributed</groupId>
-            <artifactId>shell</artifactId>
+            <groupId>org.opennms.features.scv</groupId>
+            <artifactId>org.opennms.features.scv.api</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+            <version>2.8.0</version>
+        </dependency>
     </dependencies>
-
 </project>
