+++ features/minion/core/pom.xml
@@ -15,6 +15,6 @@
         <module>impl</module>
         <module>features</module>
         <module>repository</module>
-        <module>minion-identity-impl</module>
+        <module>minion-rest-client</module>
     </modules>
 </project>
