+++ features/scv/dominion-grpc-impl/pom.xml
@@ -11,7 +11,7 @@
     <groupId>org.opennms.features.scv</groupId>
     <artifactId>org.opennms.features.scv.dominion-grpc-impl</artifactId>
     <packaging>bundle</packaging>
-    <name>OpenNMS :: Features :: SCV :: Dominion gRPC Impl.</name>
+    <name>OpenNMS :: Features :: SCV :: Dominion gRPC Impl</name>
     <build>
         <plugins>
             <plugin>
@@ -60,6 +60,12 @@
         <dependency>
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-library</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.opennms.features.distributed</groupId>
+            <artifactId>core-api</artifactId>
+            <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
