+++ features/scv/dominion-grpc-impl/src/test/java/org/opennms/dominion/features/scv/grpc/DominionSecureCredentialsVaultIT.java
@@ -31,7 +31,6 @@ package org.opennms.dominion.features.scv.grpc;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
-import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 import java.net.ServerSocket;
@@ -108,7 +107,7 @@ public class DominionSecureCredentialsVaultIT {
 
     @Test
     public void canGetAliases() {
-        assertThat(scv.getAliases(), contains(aliases));
+        assertThat(scv.getAliases(), contains(aliases.toArray()));
     }
 
     @Test
